Extract cluster and node collectors into separate files
